__OTHER_FRAMEWORKS__/grafana/drone-graph-main/pyproject.toml
[project]
name = "drone_graph"
version = "0.1.0"
description = "Graph your Drone pipelines"
readme = "README.md"
requires-python = ">=3.13"
license = "AGPL-3.0-only"
dependencies = [
    "graphviz>=0.20.3",
    "pyyaml>=6.0.2",
]

[project.scripts]
drone-graph = "drone_graph.__main__:main"
